backend/requirements.txt
Django==2.1.15
django-cors-headers==3.3.0
djangorestframework==3.11.0
pytz==2020.1
